+++ src/Frame.cpp
@@ -16,16 +16,25 @@
 
 namespace SLAM_demo {
 
+using std::vector;
 using cv::Mat;
 
 unsigned Frame::nNextIdx = 0;
+
+const int Frame::NUM_BLK_X = 3;
+const int Frame::NUM_BLK_Y = 3;
 
 Frame::Frame(const cv::Mat& img, double timestamp) :
     mTimestamp(timestamp), mnIdx(nNextIdx++)
 {
     // configure feature extractor
     mpFeatExtractor = cv::ORB::create(
-        Config::nFeatures(), Config::scaleFactor(), Config::nLevels());
+        Config::nFeatures() / (NUM_BLK_X * NUM_BLK_Y),
+        Config::scaleFactor(), Config::nLevels(),
+        7, // edgeThreshold
+        0, 2, cv::ORB::HARRIS_SCORE,
+        31, // patchSize
+        20);
     extractFeatures(img);
 }
 
@@ -57,7 +66,45 @@ cv::Mat Frame::coordCam2Img(const cv::Mat& Xc) const
 
 void Frame::extractFeatures(const cv::Mat& img)
 {
-    mpFeatExtractor->detectAndCompute(img, cv::noArray(), mvKpts, mDescs);
+    //mpFeatExtractor->detectAndCompute(img, cv::noArray(), mvKpts, mDescs);
+    // block-based feature extraction
+    mvKpts.reserve(Config::nFeatures());
+    int nImgWidth = img.cols;
+    int nImgHeight = img.rows;
+    int nBlkWidth = nImgWidth / NUM_BLK_X;
+    int nBlkHeight = nImgHeight / NUM_BLK_Y;
+    for (int i = 0; i < NUM_BLK_Y; ++i) {
+        int nBlkTLY = nBlkHeight * i; // top-left y coord of a block
+        if (i == NUM_BLK_Y - 1) {
+            nBlkHeight = nImgHeight - (NUM_BLK_Y - 1) * nBlkHeight;
+        }
+        for (int j = 0; j < NUM_BLK_X; ++j) {
+            int nBlkTLX = nBlkWidth * j; // top-left x coord of a block
+            if (j == NUM_BLK_X - 1) {
+                nBlkWidth = nImgWidth - (NUM_BLK_X - 1) * nBlkWidth;
+            }
+            // get image ROI of the target block
+            cv::Rect roi = cv::Rect(nBlkTLX, nBlkTLY, nBlkWidth, nBlkHeight);
+            Mat imgROI = Mat(img, roi);
+            // feature extraction
+            vector<cv::KeyPoint> vKpts;
+            Mat descs;
+            mpFeatExtractor->detectAndCompute(
+                imgROI, cv::noArray(), vKpts, descs);
+            // map local keypoint coords to global image coords
+            for (cv::KeyPoint& kpt : vKpts) {
+                kpt.pt.x += nBlkTLX;
+                kpt.pt.y += nBlkTLY;
+            }
+            // append features in the ROI to global containers
+            mvKpts.insert(std::end(mvKpts), std::begin(vKpts), std::end(vKpts));
+            if (i == 0 && j == 0) {
+                descs.copyTo(mDescs);
+            } else {
+                mDescs.push_back(descs.clone());
+            }
+        }
+    }
     // undistort keypoint coordinates
     undistortKpts();
  }